Create MaxCounters.java
MaxCounters problem on Codility. This code achieves a score of 100 percent on both correctness and performance (O(M+N)). 

The key idea here is to make sure that we check only for the maximum value up to the last time N+1 occurs in A, then from that point on, start updating the counter accordingly.